fix: avoid signaling reused PGID after leader reap
After Wait reaps the leader, groupRunnable treats a live process at
pid==pgid as PGID reuse so drainGroup does not TERM/KILL an unrelated
group. Split containment unit tests under the 300-line test-file guideline.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
